https://jira.catena-x.net/browse/KA-59 unit tests for data plane, move text-based parameter replacement to query processor to prepare real syntax replacement.
