+++ dataspace/agents/remoting/pom.xml
@@ -97,13 +97,6 @@
           <type>pom</type>
           <scope>import</scope>
         </dependency>
-        <dependency>
-            <groupId>org.junit</groupId>
-            <artifactId>junit-bom</artifactId>
-            <version>5.9.0</version>
-            <type>pom</type>
-            <scope>import</scope>
-        </dependency>
     </dependencies>
     </dependencyManagement>
