+++ ontology/tools/pom.xml
@@ -47,7 +47,6 @@
        <dependency>
          <groupId>org.junit.jupiter</groupId>
          <artifactId>junit-jupiter-engine</artifactId>
-         <version>5.5.2</version>
          <scope>test</scope>
        </dependency>
