+++ pom.xml
@@ -39,6 +39,13 @@
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
+            <dependency>
+             <groupId>org.junit</groupId>
+             <artifactId>junit-bom</artifactId>
+             <version>5.9.0</version>
+             <type>pom</type>
+             <scope>import</scope>
+            </dependency>
       </dependencies>
     </dependencyManagement>
     <repositories>
